update common to reflect the new module name, some fixes for session, and make parsing the moduli file only return longs.
git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@4343 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb

+++ twisted/conch/ssh/primes.py
@@ -33,7 +33,7 @@ def parseModuliFile(filename):
             continue
         tim, typ, tst, tri, size, gen, mod = l.split()
         size = int(size) + 1
-        gen = int(gen)
+        gen = long(gen)
         mod = long(mod, 16)
         if not primes.has_key(size):
             primes[size] = []
